Use size_t instead of uint32_t ad/write count-args.

+++ cpukit/libcsupport/include/rtems/libio.h
@@ -67,13 +67,13 @@ typedef int (*rtems_filesystem_close_t)(
 typedef ssize_t (*rtems_filesystem_read_t)(
   rtems_libio_t *iop,
   void          *buffer,
-  uint32_t       count
+  size_t         count
 );
 
 typedef ssize_t (*rtems_filesystem_write_t)(
   rtems_libio_t *iop,
   const void    *buffer,
-  uint32_t      count
+  size_t         count
 );
 
 typedef int (*rtems_filesystem_ioctl_t)(
